Refactor Header and SideNav for easier packaging as a static route
Refactor the Header and SideNav components so that they can be more
easily consumed as part of a statically defined route using the new
React Router data APIs. This means the header and logout button are
no longer pure UI components, and are moved to the containers folder
instead, so they can use Dashboard-specific APIs and data as needed.
These components are quite specific to the Dashboard application
anyway so didn't really make sense as pure UI components in the first
place.
