+++ src/containers/Header/Header.scss
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2023 The Tekton Authors
+Copyright 2019-2024 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -10,5 +10,18 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-/* istanbul ignore file */
-export { default } from './Header';
+
+header.#{$prefix}--header.tkn--header {
+  .#{$prefix}--list-box__wrapper {
+    margin-block-start: 0.5rem;
+    margin-inline-end: 1rem;
+
+    &:last-child {
+      margin-inline-end: 2rem;
+    }
+  }
+
+  .tkn--logout-btn {
+    padding-inline-end: 2px;
+  }
+}
