+++ src/containers/index.js
@@ -25,12 +25,14 @@ export { default as CustomRun } from './CustomRun';
 export { default as CustomRuns } from './CustomRuns';
 export { default as EventListener } from './EventListener';
 export { default as EventListeners } from './EventListeners';
+export { default as Header } from './Header';
 export { default as HeaderBarContent } from './HeaderBarContent';
 export { default as ImportResources } from './ImportResources';
 export { default as Interceptors } from './Interceptors';
 export { default as LabelFilter } from './LabelFilter';
 export { default as ListPageLayout } from './ListPageLayout';
 export { default as LoadingShell } from './LoadingShell';
+export { default as LogoutButton } from './LogoutButton';
 export { default as NamespacedRoute } from './NamespacedRoute';
 export { default as NamespacesDropdown } from './NamespacesDropdown';
 export { default as NotFound } from './NotFound';
